requirements.txt
flask
flask-cors
python-dotenv
requests
edge-tts
nest_asyncio
SpeechRecognition
deep-translator
pydub
pg8000
